Refactor & expand tests + fix bug
