+++ src/hooks/models/attendees/useIsAttendeeDisabledCallback.js
@@ -51,6 +51,9 @@ export const useIsAttendeeDisabledCallback = (session, attendees) => {
     session?.dayNumber,
     attendees
   )
+
+  // if there is a pending session, that means it's waiting for a submission response,
+  // so we don't want any attendees to be bookable during that time
   const pendingSession = useStoreItems(CATEGORIES.PENDING_SESSION)
 
   return useCallback(
